Refactor: Move components and consolidate ui components.
