+++ src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export * from './section';
+export * from './container';
